ダイナミックルートを採用
ジャンルデータにダイナミックルートを採用
ジャンルデータの下層ジャンルIDからリンクを設置
